# openapi.yaml
openapi: 3.1.0
info:
  title: Local Qwen LoRA API
  description: A local API for serving Qwen models with LoRA adapters, mimicking the OpenAI Chat Completion format.
  version: 1.0.0
servers:
  - url: http://localhost:11434
    description: Local development server
paths:
  /v1/chat/completions:
    post:
      summary: Create chat completion
      operationId: chat_completions_v1_chat_completions_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPException'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'

components:
  schemas:
    Message:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          title: Role
          description: The role of the messages author (e.g., user, assistant, system).
        content:
          type: string
          title: Content
          description: The contents of the message.

    ChatCompletionRequest:
      type: object
      required:
        - messages
      properties:
        model:
          type: string
          title: Model
          default: qwen-lora
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
          title: Messages
        max_tokens:
          type: integer
          title: Max Tokens
          default: 512
          nullable: true
        temperature:
          type: number
          title: Temperature
          default: 0.7
          nullable: true

    ChatCompletionResponse:
      type: object
      properties:
        id:
          type: string
          title: ID
        object:
          type: string
          title: Object
          example: chat.completion
        created:
          type: integer
          title: Created
        model:
          type: string
          title: Model
        choices:
          type: array
          items:
            $ref: '#/components/schemas/Choice'
        usage:
          $ref: '#/components/schemas/Usage'

    Choice:
      type: object
      properties:
        index:
          type: integer
        message:
          $ref: '#/components/schemas/Message'
        finish_reason:
          type: string
          example: stop

    Usage:
      type: object
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer

    HTTPException:
      type: object
      properties:
        detail:
          type: string
          title: Detail

    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
    
    ValidationError:
      type: object
      required:
        - loc
        - msg
        - type
      properties:
        loc:
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type